backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
google-generativeai
google-genai
volcengine-python-sdk
volcengine-python-sdk[ark]
openai
dashscope
zai-sdk
voyageai
edge-tts
python-jose[cryptography]
PyJWT
passlib[bcrypt]
python-dotenv
pandas
requests
httpx
python-multipart
email-validator
PyMuPDF
seaborn
scikit-learn
json-repair
zhconv
boto3
